Add close confirmation and configuration loading improvements
This change introduces a close confirmation dialog when attempting to close the application. This change also adds checks to the configuration reading operations, and in case of failure to read the configuration, an error message will be displayed to the user. Various code improvements for better validation and thread safety were also included.