Add tabbed pane and refactor UI setup
Implemented a tabbed pane into the content panel for better organization of UI elements. The previous elements, a JLabel and JButton, are now placed under one of the tabs. Also, the methods for setting up the UI and third-party look & feel libraries have been refactored for readability and extensibility. System specific UI setups have been provided for macOS, Linux and a placeholder for Windows.