Implement concise configuration rendering
The code has been modified to use ConfigRenderOptions for generating a concise and formatted config string. This replaces the prior method of rendering the Config object to a ConfigDocument which was then written to a file. The new solution fixes a bug where new comments where written every time the file was saved